pyproject.toml
[tool.poetry]
name = "personal-assistant"
version = "0.1.0"
description = ""
authors = ["Valentina Rudenko <valentino4ka2009@gmail.com>, Serhii Rezvaniuk <o_serg@ukr.net>, Yuliia Diakiv <uliadiakiv@gmail.com>, Ksenia Kosarevska <kosarevskayakseniia@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
cloudinary = "^1.40.0"
django-cloudinary-storage = "^0.3.0"
alembic = "^1.13.2"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
